+++ Tests/APITests/TestHelpers.swift
@@ -130,6 +130,7 @@ private func registerBaseTestMigrations(on app: Application) {
     app.migrations.add(AddTestSetupLastRetestedManifestHash())
     app.migrations.add(CreatePreEnrollments())
     app.migrations.add(AddUserLastSeenAt())
+    app.migrations.add(AddBrightSpaceSyncFields())
 }
 
 private func registerObservabilityTestMigrations(on app: Application) {